Added copy optimization to basic version
